@@ -37,6 +37,8 @@ dependencies:
   firebase_database: ^10.4.0
   firebase_auth: ^4.16.0
 
+  english_words: ^4.0.0
+
   cupertino_icons: ^1.0.6
 
   image_picker: ^1.0.7
@@ -71,11 +73,13 @@ dependencies:
 
   chewie: ^1.7.5
 
-  font_awesome_flutter: ^9.2.0
+  font_awesome_flutter: ^10.7.0
 
   uuid: ^4.3.3
 
-  chips_choice: ^3.0.1
+  chips_choice: ^3.0.1  get: ^4.6.6
+
+
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -119,7 +123,11 @@ flutter:
   assets:
     - assets/blank_pfp.png
     - assets/base_image.png
-
+    - assets/default_profile_photo.png
+    - assets/profile_photo_1.png
+    - assets/profile_photo_2.png
+    - assets/profile_photo_3.png
+    
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
 
