@@ -4,17 +4,21 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:intl/intl.dart';
 import 'package:info_hub_app/threads/thread_replies.dart';
 import 'package:flutter/cupertino.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:info_hub_app/threads/name_generator.dart';
 
 class CustomCard extends StatefulWidget {
   final QuerySnapshot? snapshot;
   final int index;
   final FirebaseFirestore firestore;
+  final FirebaseAuth auth;
 
   const CustomCard({
     Key? key,
     this.snapshot,
     required this.index,
     required this.firestore,
+    required this.auth,
   }) : super(key: key);
 
   @override
@@ -24,7 +28,7 @@ class CustomCard extends StatefulWidget {
 class _CustomCardState extends State<CustomCard> {
   late TextEditingController titleInputController;
   late TextEditingController descriptionInputController;
-  late TextEditingController nameInputController;
+  //late TextEditingController nameInputController;
 
   @override
   void initState() {
@@ -37,14 +41,14 @@ class _CustomCardState extends State<CustomCard> {
     nameInputController = TextEditingController(text: docData['name']); */
     titleInputController = TextEditingController();
     descriptionInputController = TextEditingController();
-    nameInputController = TextEditingController();
+    //nameInputController = TextEditingController();
   }
 
   @override
   void dispose() {
     titleInputController.dispose();
     descriptionInputController.dispose();
-    nameInputController.dispose();
+    //nameInputController.dispose();
     super.dispose();
   }
 
@@ -56,15 +60,17 @@ class _CustomCardState extends State<CustomCard> {
 
     var title = docData['title'] ?? 'No title';
     var description = docData['description'] ?? 'No description';
-    var name = docData['name'] ?? 'Unknown';
+    var creator = docData['creator'] ?? 'Unknown';
+    var currentUserId = widget.auth.currentUser!.uid;
+    //var name = docData[''] ?? 'Unknown';
     var timestamp = docData['timestamp']?.toDate();
     var formatter = timestamp != null
         ? DateFormat("dd-MMM-yyyy 'at' HH:mm").format(timestamp)
         : 'Timestamp not available';
+    var authorName = generateUniqueName(creator);
 
     return Column(
       children: <Widget>[
-        //Container(
         SizedBox(
           height: 140,
           child: Card(
@@ -80,10 +86,14 @@ class _CustomCardState extends State<CustomCard> {
                     child: Material(
                       color: Colors.transparent,
                       child: InkWell(
+                        key: Key('navigateToThreadReplies_${widget.index}'),
                         onTap: () {
                           Navigator.of(context).push(CupertinoPageRoute(
-                              builder: (BuildContext context) =>
-                                  const ThreadReplies())); // Replace with actual thread replies page
+                              builder: (BuildContext context) => ThreadReplies(
+                                    threadId: docId,
+                                    firestore: widget.firestore,
+                                    auth: widget.auth,
+                                  )));
                         },
                         child: Text(
                           title,
@@ -98,22 +108,39 @@ class _CustomCardState extends State<CustomCard> {
                     ),
                   ),
                   const SizedBox(width: 8),
-                  IconButton(
-                    icon: const Icon(FontAwesomeIcons.edit, size: 15),
-                    onPressed: () {
-                      _showDialog(context, docId);
-                    },
-                  ),
-                  IconButton(
-                    icon: const Icon(FontAwesomeIcons.trashAlt, size: 15),
-                    onPressed: () async {
-                      if (!mounted) return;
-                      await widget.firestore
-                          .collection("thread")
-                          .doc(docId)
-                          .delete();
-                    },
-                  ),
+                  if (currentUserId == creator)
+                    IconButton(
+                      key: Key('editButton_$docId'),
+                      icon: const Icon(FontAwesomeIcons.edit, size: 15),
+                      onPressed: () {
+                        _showDialog(context, docId);
+                      },
+                    ),
+                  if (currentUserId == creator)
+                    IconButton(
+                      icon: const Icon(FontAwesomeIcons.trashAlt, size: 15),
+                      onPressed: () async {
+                        if (!mounted) return;
+                        // batch delete removes up to 500 replies at a time from the replies collection due to firestore limitations
+                        final replyQuerySnapshot = await widget.firestore
+                            .collection("replies")
+                            .where('threadId', isEqualTo: docId)
+                            .get();
+
+                        final WriteBatch batch = widget.firestore.batch();
+                        for (DocumentSnapshot replyDoc
+                            in replyQuerySnapshot.docs) {
+                          batch.delete(replyDoc.reference);
+                        }
+
+                        await batch.commit();
+
+                        await widget.firestore
+                            .collection("thread")
+                            .doc(docId)
+                            .delete();
+                      },
+                    ),
                 ],
               ),
               subtitle: Column(
@@ -131,7 +158,7 @@ class _CustomCardState extends State<CustomCard> {
                     mainAxisAlignment: MainAxisAlignment.start,
                     children: <Widget>[
                       Text(
-                        "$name: ",
+                        "$authorName: ",
                         style: const TextStyle(
                           fontSize: 14,
                           color: Color.fromARGB(255, 255, 0, 0),
@@ -173,10 +200,9 @@ class _CustomCardState extends State<CustomCard> {
 
     if (!mounted) return;
 
-    // Reinitialize the TextEditingController instances with the latest document data
     titleInputController.text = docData['title'] ?? '';
     descriptionInputController.text = docData['description'] ?? '';
-    nameInputController.text = docData['name'] ?? '';
+    //nameInputController.text = docData['name'] ?? '';
 
     await showDialog(
       context: context,
@@ -190,7 +216,7 @@ class _CustomCardState extends State<CustomCard> {
                   mainAxisSize: MainAxisSize.min,
                   children: <Widget>[
                     const Text("Please fill out the form"),
-                    TextField(
+                    /*TextField(
                       key: const Key('Author'),
                       autofocus: true,
                       autocorrect: true,
@@ -200,7 +226,7 @@ class _CustomCardState extends State<CustomCard> {
                             showErrorName ? "Please enter your name" : null,
                       ),
                       controller: nameInputController,
-                    ),
+                    ), */
                     TextField(
                       key: const Key('Title'),
                       autofocus: true,
@@ -240,7 +266,7 @@ class _CustomCardState extends State<CustomCard> {
                 TextButton(
                   onPressed: () {
                     setState(() {
-                      showErrorName = nameInputController.text.isEmpty;
+                      //showErrorName = nameInputController.text.isEmpty;
                       showErrorTitle = titleInputController.text.isEmpty;
                       showErrorDescription =
                           descriptionInputController.text.isEmpty;
@@ -250,9 +276,9 @@ class _CustomCardState extends State<CustomCard> {
                         !showErrorTitle &&
                         !showErrorDescription) {
                       widget.firestore.collection("thread").doc(docId).update({
-                        "name": nameInputController.text,
                         "title": titleInputController.text,
-                        "description": descriptionInputController.text,
+                        "description":
+                            "${descriptionInputController.text} (edited)",
                         "timestamp": FieldValue.serverTimestamp(),
                       }).then((response) {
                         //print(response.id);
