@@ -199,16 +199,17 @@ class _RegistrationScreenState extends State<RegistrationScreen> {
 
                     if (userModel != null) {
                       // Registration was successful, navigate to the main application page
-                      Navigator.pushReplacement(
-                        context,
-                        MaterialPageRoute(
-                          builder: (context) => Base(
+                      Widget nextPage = Base(
                             auth: widget.auth,
                             storage: widget.storage,
                             firestore: widget.firestore,
-                          ),
-                        ),
+                          );
+                      Navigator.pushAndRemoveUntil(
+                        context,
+                        MaterialPageRoute(builder: (context) => nextPage),
+                        (Route<dynamic> route) => false,
                       );
+
                     } else {
                       // Show error message if anything goes wrong in the auth process
                       ScaffoldMessenger.of(context).showSnackBar(
